// src/lib/pocketbase-types.ts
export const Collections = {
	Authorigins: "_authOrigins",
	Externalauths: "_externalAuths",
	Mfas: "_mfas",
	Otps: "_otps",
	Superusers: "_superusers",
	Answers: "answers",
	Questions: "questions",
	Users: "users",
} as const
export type Collections = typeof Collections[keyof typeof Collections]

// Alias types for improved usability
export type IsoDateString = string
export type IsoAutoDateString = string & { readonly autodate: unique symbol }
export type RecordIdString = string
export type FileNameString = string & { readonly filename: unique symbol }
export type HTMLString = string

type ExpandType<T> = unknown extends T
	? T extends unknown
		? { expand?: unknown }
		: { expand: T }
	: { expand: T }

// System fields
export type BaseSystemFields<T = unknown> = {
	id: RecordIdString
	collectionId: string
	collectionName: Collections
} & ExpandType<T>
// ...
export type AnswersRecord = {
	content: string
	created: IsoAutoDateString
	id: string
	question?: RecordIdString
	title: string
	updated: IsoAutoDateString
}

export const QuestionsCategoryOptions = {
	"ESE124 Related": "ESE124 Related",
	"ETC": "ETC",
} as const
export type QuestionsCategoryOptions = typeof QuestionsCategoryOptions[keyof typeof QuestionsCategoryOptions]
export type QuestionsRecord = {
	category?: QuestionsCategoryOptions
	content: string
	created: IsoAutoDateString
	id: string
	title: string
	updated: IsoAutoDateString
}
// ...
export type AnswersResponse<Texpand = unknown> = Required<AnswersRecord> & BaseSystemFields<Texpand>
export type QuestionsResponse<Texpand = unknown> = Required<QuestionsRecord> & BaseSystemFields<Texpand>